requirements.txt
python-dotenv==0.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
PyDrive==1.3.1
vk-api==11.4.0
python-telegram-bot==11.1.0
requests==2.21.0
PySocks==1.6.8
beautifulsoup4==4.7.1
lxml==4.3.3
urlextract==0.10